deleated src/features and relocated to src/data
